Begin refactoring the SkyPoint class.
We want to simplify the SkyPoint class and remove all of the
computation methods. We're going to do this by converting them into
functions in the namespace KSEngine::OldPointFunctions.

Eventually, we'll be able to remove them all, but for now we're going
to move them out, so that we can work on them seperately.

This commit moves the updateCoords function out of the SkyPoint.
Unfortunately, updating the coordinates is done in three different
ways using virtual methods, so we can't just remove it. Instead, we
have to settle for removing the bulk of the computation.