Move the GR calculations out of the skypoint.
Note: this one is tricky, because these calculations need to
know the position of the sun. I don't want these functions to have
any references to state not passed in as arguments, so I've cheated
and just put the KSSun pointer into the KSNumbers, which is already so
big and bloated it can't really get a great deal worse.

More importantly, we'd like to be able to get rid of the KSNumbers
class in its entirety, so this serves the purpose of reminding us to
sort out this issue when we decouple that part of the code.

CCMAIL: akarshsimha@gmail.com