Move the deprecess() function.
Note: this function really confuses two seperate issues:
first, deprecession, and second, fixing the catalog coordinates for
the point.

This causes problems, first because the function is now stateful, second
because it means that it's never explicit when the catalog coordinates
are set, and third because the implementation is incorrect. If a point has
invalid catalog coordinates, then it will update them -- but we might not
be deprecessing to J2000. We might be deprecessing to B1950 or any other
epoch.