Move the remaining precession functions.
The implementation of the precessFromAnyEpoch function is really
quite bizarre, and when we write a new implementation, we should
not just translate the existing one, but start over.

This commit also makes the datahandlers depend on Eigen, since
they need to use the KSEngine stuff, which uses Eigen.
It's unclear why KStars does anything at all when Eigen is not found
(see the root CMakeLists.txt).

CCMAIL: akarshsimha@gmail.com