Try to move some of the nutation calculations from KSNumbers.
This implementation doesn't give the same result as of yet, but
it's not obvious (to me) why. It's also
not clear what the difference between the two obliquity calculations
in KSNumbers is, but it's quite large and I presume important. The ported
code follows Meeus, chapter 20.

If anyone wants to take a brief look and see whether I've made
an obvious mistake, it'd be appreciated.

CCMAIL: akarshsimha@gmail.com
CCMAIL: ra.rishab@gmail.com