Add an implementation of nutation.
Note that we use the exact method always, unlike in the
old implementation, where we use an approximation. However,
even the exact method here should be much faster than the
trig in the old method.

Note also that this implementation doesn't have a test yet, since
it's getting its numbers from a function which is known to give
incorrect results. When the AstroVars tests pass, then we
should add a nutation test.

CCMAIL: akarshsimha@gmail.com