Change from using quaternions to matrices.
This is rather unfortunate, but unfortunately we need to
be using orthogonal matrices instead of quaternions.
The reason is that the conversion from equatorial to horizontal
coordinates has a sign change, with the result that the determinant
is -1 instead of +1. This means that that map cannot be represented in
the usual way by quaternions, so we need to use matrices instead.

CCMAIL: akarshsimha@gmail.com