Add an implementation of aberration.
This should be used as a stopgap until we have a better method.
The problem is that aberration isn't quite like the other transforms
we do in terms of being a rotation of 3d space. However, it is apparently
a Lorentz transform, so I think that it should be possible to do
things linear-algebraically? I'm not sure.

There should really be a de-aberrate function, as well.

Note also that we do the aberration conversion in ecliptic coordinates for
clarity of implementation. Use the appropriate functions to change coordinates
first.