Apply coordinate conversions in place.
To do this we move the function from the context to the buffer class.
This also means that KSClBuffer and KSClContext become more intertwined.
I think that this is OK, since the external interface remains clear,
it's just that internally they share resources. For instance, the
context should keep a list of all of the kernels, and the buffer needs
this list, but outside of KSCl nobody needs these, and we don't
want to expose those details (or anything in namespace cl).