Remove proto-bindings for OpenCL details.
I've decided that we should try to quarantine the OpenCL
code to the smallest possible area. Not exposing OpenCL
to the rest of KStars means that in the event we decide to
back out of using OpenCL and just use OpenMP or something,
it's significantly easier to do so.

CCMAIL:akarshsimha@gmail.com