Carry out code reorganization.
The plan is to put all of the new, clean code in ksengine/
and have it not depend on anything else in KStars, so that we
can make some progress on modularizing the code.

The old code that was pulled out of SkyPoint is in kstars/oldengine/
and depends on all kinds of random stuff.