Add a datacomponents tree.
The idea here is to replace the SkyComponent hierarchy, which
currently complects:

*   Data loading / storage
*   Keeping data up-to-date with the current simulation time
*   Drawing the data on the screen

and having a new heirarchy which is solely responsible for
storing and maintaining the data.

The code for drawing will be moved out into a seperate class,
so that our code is more in line with an "ideal MVC" setup.

The data components will have an update() function, but unlike in
the SkyComponents hierarchy, they are responsible only for getting
the current *equatorial* coordinates for the given time. The reason
for this is that the horizontal coordinates are dependent on the
observer, and are thus more appropriately placed into the "view"
part of the MVC setup.

CCMAIL: akarshsimha@gmail.com