Have DataComponent compute horizontal coords
I think that it would be better to have it compute equatorial
coordinates, and leave the horizontal coordinate calculations to be done
by the view, but this makes things slightly awkward given the way that
we've set up the skycomponent heirarchy, since, for example,
the horizontal coordinate grid, and the horizon are given in horizontal
coordinates. However, I think that for now it's best to try to leave
things mostly the way that they are already, so I don't want to try to
come up with a totally new design for these things.

CCMAIL:akarshsimha@gmail.com