Add devcontainer support (for a reproducable developer build environment)
This adds a .devcontainer directory (https://containers.dev/).  The three
small files in this directory create a developer environment that runs inside
of a docker container.  Most modern IDEs include support for devcontainers, but
I tested with vscode.

Usage of devcontainers is entirely optional and for a non devcontainer based
dev environment nothing changes.  If on an IDE with devcontainer support the user
will be asked if they want to develop inside a container.

The container is fedora based and installs all necessary build environment files
(revision locked to fedora:42 for repeatability).  It also downloads and builds the correct
version of the solver library (needed to keep working with qt5/kstars)).

After the container is running, jsut type "make -j16; sudo make install" to build and install.